requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Deprecated==1.2.18
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.23
stripe==7.8.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.0.1
wrapt==1.17.3
